% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caratMULT-package.R
\docType{package}
\name{caratMULT-package}
\alias{caratMULT}
\alias{caratMULT-package}
\title{caratMULT: Analysis Methods for Covariate-Adaptive Randomized Trials with Multiple Treatments}
\description{
This package implements all of the estimators considered in Gu et al. (2023) for estimating and inferring the multiple treatment effect under covariate-adaptive randomization.
}
\keyword{internal}
