% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tau.ss}
\alias{tau.ss}
\title{Stratum-specific estimator}
\usage{
tau.ss(Y, X, A, B, a, b, conf.level = 0.95)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same as the number of subjects.}

\item{X}{a numeric design matrix containing additional covariates used in the model.}

\item{A}{a numeric vector of treatment assignments. The assignments should be consecutive integers from 0 to nA, where 
nA is the number of different treatments (including control group).}

\item{B}{a numeric vector of stratum labels. Its length should be the same as the number of subjects.}

\item{a, b}{an integer denote the treatment in \code{A}. The result will present the estimated treatment effect between selected
treatment \code{a} and treatment \code{b}}

\item{conf.level}{confidence level of the interval. Default is 0.95.}
}
\value{
A list of class \code{"htest"} containing the following components:
\item{statistic}{the value of the t-statistic.} \item{p.value}{the p-value for
the test.} \item{conf.int}{a confidence interval under chosen level
\code{conf.level} for the difference in treatment effect between treatment
group and control group.} \item{estimate}{estimated treatment effect
difference between treatment group and control group.} \item{method}{a
character string indicating what type of regression was performed.}
}
\description{
Estimating and inferring the treatment effect based the on the stratum-specific estimator.
}
\details{
Estimating and inferring the treatment effect based on the stratum-specific estimator. It implements the 
methods as described in Section 3.3, Gu et al.(2023).
}
\examples{
#The code replicates the simulation setting of Model 1 in Section 4, Gu et al. (2023).

nA = 3
nS = 4
N=600
p_s = c(2,2,3,3)/10

pi_a = rep(c(1,1,1)/3)

p=1
beta0 <- matrix(rep(1,p),p,1)
beta1 <- matrix(rep(2,p),p,1)
beta2 <- matrix(rep(3,p),p,1)

X1 <- sample(1:nS,N,prob = p_s,replace = TRUE)
#additional variable
X2 <- matrix(rnorm(N*p,0,1),N,p)

trt <- sample(0:(nA-1),N,replace=TRUE,pi_a)

ind0 <- which(trt==0)
ind1 <- which(trt==1)
ind2 <- which(trt==2)


Y0 <- 0 + X1[ind0] + X2[ind0,]\%*\%beta0
Y1 <- 1 + X1[ind1] + X2[ind1,]\%*\%beta1
Y2 <- 2 + X1[ind2] + X2[ind2,]\%*\%beta2
Y <- rep(0,N)
Y[ind0] <- Y0
Y[ind1] <- Y1
Y[ind2] <- Y2
Y <- Y + rnorm(N,0,1)
tau.ss(Y,X2,trt,X1,1,0)
}
\references{
Gu Y., Liu H., Ma W.(2023). \emph{Regression-based multiple treatment effect estimation under covariate-adaptive 
randomization.} In review.
}
